Refactoring TrasnformationLogic
