+++ srcs/cli/filters/__init__.py
@@ -1,0 +1,4 @@
+"""
+Filter modules for image transformation pipeline.
+Each filter is responsible for its specific transformation logic.
+"""
